DGCF completed and tested.

+++ external/models/deepconn/DeepCoNN.py
@@ -149,7 +149,7 @@ class DeepCoNN(RecMixin, BaseRecommenderModel):
                 stop_batch = min(start_batch + self._batch_eval, self._num_items)
                 item_reviews = list(
                     itemgetter(*list(range(start_batch, stop_batch)))(items_tokens))
-                out_items[start_batch: stop_batch] = self._model.conv_items(tf.Variable(item_reviews, dtype=tf.float32))
+                out_items[start_batch: stop_batch] = self._model.conv_items(tf.Variable(item_reviews, dtype=tf.int32))
                 t.update()
         self.logger.info('Convolutions for all items is complete!')
